requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
Django==4.0.2
django-filter==21.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.2.0
flake8==3.9.2
flake8-django==1.1.2
iniconfig==1.1.1
Markdown==3.3.6
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2021.5
